// src/logic/rest/api.ts
export interface RoomMessage {
    /**
     * The id of the message
     * @type {number}
     * @memberof RoomMessage
     */
    id: number;
    /**
     * The senderId of the message
     * @type {number}
     * @memberof RoomMessage
     */
    senderId: number;
    /**
     * The roomId of the message
     * @type {number}
     * @memberof RoomMessage
     */
    roomId: number;
    /**
     * The username of the sender
     * @type {string}
     * @memberof RoomMessage
     */
    senderName: string;
    /**
     * The actual message
     * @type {string}
     * @memberof RoomMessage
     */
    content: string;
    /**
     * The timestamp of the message
     * @type {Date}
     * @memberof RoomMessage
     */
    timeSent: Date;
}